Factor tt::has_equivalence to its own file
